sort keyframes, create the main buffer (only int16_t, so only 16 bit depth audio as of now)

+++ include/files/extract.h
@@ -1,4 +1,5 @@
 #include <stdbool.h>
+#include <stdint.h>
 
 #ifndef EXTRACT_H
 #define EXTRACT_H
@@ -6,7 +7,7 @@
 #include "extract.h"
 
 char *parseLineValueS(char *variable, char *path);
-int parseLineValueI(char *variable, char *path);
+int64_t parseLineValueI(char *variable, char *path);
 float parseLineValueF(char *variable, char *path);
 bool parseLineValueB(char *variable, char *path);
